+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -927,7 +927,7 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
             result
         }
         Function::Choosecols | Function::Chooserows => vec![Signature::Vector; arg_count],
-        Function::Tocol | Function::Torow => {
+        Function::Tocol | Function::Torow | Function::Expand => {
             let mut result = vec![Signature::Vector; arg_count];
             for item in result.iter_mut().skip(1) {
                 *item = Signature::Scalar;
@@ -1535,6 +1535,7 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Chooserows => not_implemented(args),
         Function::Tocol => not_implemented(args),
         Function::Torow => not_implemented(args),
+        Function::Expand => not_implemented(args),
         Function::Aggregate => not_implemented(args),
         Function::Rand => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),