+++ crates/ironcalc-base/src/functions/mod.rs
@@ -544,6 +544,7 @@ pub enum Function {
     Chooserows,
     Tocol,
     Torow,
+    Expand,
 }
 
 macro_rules! impl_function_lookup {
@@ -1062,6 +1063,7 @@ impl_function_lookup! {
     chooserows      => Chooserows,
     tocol           => Tocol,
     torow           => Torow,
+    expand          => Expand,
 }
 
 pub fn english_function_names() -> Vec<String> {
@@ -1545,9 +1547,10 @@ impl Function {
             Function::Chooserows => functions.chooserows.clone(),
             Function::Tocol => functions.tocol.clone(),
             Function::Torow => functions.torow.clone(),
+            Function::Expand => functions.expand.clone(),
         }
     }
-    pub fn into_iter() -> IntoIter<Function, 467> {
+    pub fn into_iter() -> IntoIter<Function, 468> {
         [
             Function::And,
             Function::False,
@@ -2016,6 +2019,7 @@ impl Function {
             Function::Chooserows,
             Function::Tocol,
             Function::Torow,
+            Function::Expand,
         ]
         .into_iter()
     }
@@ -2074,6 +2078,7 @@ impl Function {
             Function::Chooserows => "_xlfn.CHOOSEROWS".to_string(),
             Function::Tocol => "_xlfn.TOCOL".to_string(),
             Function::Torow => "_xlfn.TOROW".to_string(),
+            Function::Expand => "_xlfn.EXPAND".to_string(),
             Function::Isformula => "_xlfn.ISFORMULA".to_string(),
             Function::Sheet => "_xlfn.SHEET".to_string(),
             Function::Formulatext => "_xlfn.FORMULATEXT".to_string(),
@@ -2685,6 +2690,7 @@ impl<'a> Model<'a> {
             Function::Chooserows => self.fn_chooserows(args, cell),
             Function::Tocol => self.fn_tocol(args, cell),
             Function::Torow => self.fn_torow(args, cell),
+            Function::Expand => self.fn_expand(args, cell),
             Function::Gauss => self.fn_gauss(args, cell),
             Function::Harmean => self.fn_harmean(args, cell),
             Function::Kurt => self.fn_kurt(args, cell),